Rename [blueprint|Blueprint] -> [integration_blueprint|Integration blueprint] (#47)
