+++ .devcontainer/configuration.yaml
@@ -3,4 +3,4 @@ default_config:
 logger:
   default: info
   logs:
-    custom_components.blueprint: debug
+    custom_components.integration_blueprint: debug
