+++ README.md
@@ -25,15 +25,15 @@ File | Purpose
 `.github/ISSUE_TEMPLATE/issue.md` | Template for issues
 `.github/settings.yml` | Probot settings to control the repository settings.
 `.vscode/tasks.json` | Tasks for the devcontainer.
-`custom_components/blueprint/translations/*` | [Translation files.](https://developers.home-assistant.io/docs/internationalization/custom_integration)
-`custom_components/blueprint/__init__.py` | The component file for the integration.
-`custom_components/blueprint/api.py` | This is a sample API client.
-`custom_components/blueprint/binary_sensor.py` | Binary sensor platform for the integration.
-`custom_components/blueprint/config_flow.py` | Config flow file, this adds the UI configuration possibilities.
-`custom_components/blueprint/const.py` | A file to hold shared variables/constants for the entire integration.
-`custom_components/blueprint/manifest.json` | A [manifest file](https://developers.home-assistant.io/docs/en/creating_integration_manifest.html) for Home Assistant.
-`custom_components/blueprint/sensor.py` | Sensor platform for the integration.
-`custom_components/blueprint/switch.py` | Switch sensor platform for the integration.
+`custom_components/integration_blueprint/translations/*` | [Translation files.](https://developers.home-assistant.io/docs/internationalization/custom_integration)
+`custom_components/integration_blueprint/__init__.py` | The component file for the integration.
+`custom_components/integration_blueprint/api.py` | This is a sample API client.
+`custom_components/integration_blueprint/binary_sensor.py` | Binary sensor platform for the integration.
+`custom_components/integration_blueprint/config_flow.py` | Config flow file, this adds the UI configuration possibilities.
+`custom_components/integration_blueprint/const.py` | A file to hold shared variables/constants for the entire integration.
+`custom_components/integration_blueprint/manifest.json` | A [manifest file](https://developers.home-assistant.io/docs/en/creating_integration_manifest.html) for Home Assistant.
+`custom_components/integration_blueprint/sensor.py` | Sensor platform for the integration.
+`custom_components/integration_blueprint/switch.py` | Switch sensor platform for the integration.
 `CONTRIBUTING.md` | Guidelines on how to contribute.
 `example.png` | Screenshot that demonstrate how it might look in the UI.
 `info.md` | An example on a info file (used by [hacs][hacs]).
@@ -65,7 +65,7 @@ to ensure Home Assistant update the code of the python library. (example `"requi
 README content if this was a published component:
 ***
 
-# blueprint
+# integration_blueprint
 
 [![GitHub Release][releases-shield]][releases]
 [![GitHub Activity][commits-shield]][commits]
@@ -78,7 +78,7 @@ README content if this was a published component:
 [![Discord][discord-shield]][discord]
 [![Community Forum][forum-shield]][forum]
 
-_Component to integrate with [blueprint][blueprint]._
+_Component to integrate with [integration_blueprint][integration_blueprint]._
 
 **This component will set up the following platforms.**
 
@@ -94,8 +94,8 @@ Platform | Description
 
 1. Using the tool of choice open the directory (folder) for your HA configuration (where you find `configuration.yaml`).
 2. If you do not have a `custom_components` directory (folder) there, you need to create it.
-3. In the `custom_components` directory (folder) create a new folder called `blueprint`.
-4. Download _all_ the files from the `custom_components/blueprint/` directory (folder) in this repository.
+3. In the `custom_components` directory (folder) create a new folder called `integration_blueprint`.
+4. Download _all_ the files from the `custom_components/integration_blueprint/` directory (folder) in this repository.
 5. Place the files you downloaded in the new directory (folder) you created.
 6. Restart Home Assistant
 7. In the HA UI go to "Configuration" -> "Integrations" click "+" and search for "Blueprint"
@@ -103,17 +103,17 @@ Platform | Description
 Using your HA configuration directory (folder) as a starting point you should now also have this:
 
 ```text
-custom_components/blueprint/translations/en.json
-custom_components/blueprint/translations/nb.json
-custom_components/blueprint/translations/sensor.nb.json
-custom_components/blueprint/__init__.py
-custom_components/blueprint/api.py
-custom_components/blueprint/binary_sensor.py
-custom_components/blueprint/config_flow.py
-custom_components/blueprint/const.py
-custom_components/blueprint/manifest.json
-custom_components/blueprint/sensor.py
-custom_components/blueprint/switch.py
+custom_components/integration_blueprint/translations/en.json
+custom_components/integration_blueprint/translations/nb.json
+custom_components/integration_blueprint/translations/sensor.nb.json
+custom_components/integration_blueprint/__init__.py
+custom_components/integration_blueprint/api.py
+custom_components/integration_blueprint/binary_sensor.py
+custom_components/integration_blueprint/config_flow.py
+custom_components/integration_blueprint/const.py
+custom_components/integration_blueprint/manifest.json
+custom_components/integration_blueprint/sensor.py
+custom_components/integration_blueprint/switch.py
 ```
 
 ## Configuration is done in the UI
@@ -126,11 +126,11 @@ If you want to contribute to this please read the [Contribution guidelines](CONT
 
 ***
 
-[blueprint]: https://github.com/custom-components/blueprint
+[integration_blueprint]: https://github.com/custom-components/integration_blueprint
 [buymecoffee]: https://www.buymeacoffee.com/ludeeus
 [buymecoffeebadge]: https://img.shields.io/badge/buy%20me%20a%20coffee-donate-yellow.svg?style=for-the-badge
 [commits-shield]: https://img.shields.io/github/commit-activity/y/custom-components/blueprint.svg?style=for-the-badge
-[commits]: https://github.com/custom-components/blueprint/commits/master
+[commits]: https://github.com/custom-components/integration_blueprint/commits/master
 [hacs]: https://github.com/custom-components/hacs
 [hacsbadge]: https://img.shields.io/badge/HACS-Custom-orange.svg?style=for-the-badge
 [discord]: https://discord.gg/Qa5fW2R
@@ -141,4 +141,4 @@ If you want to contribute to this please read the [Contribution guidelines](CONT
 [license-shield]: https://img.shields.io/github/license/custom-components/blueprint.svg?style=for-the-badge
 [maintenance-shield]: https://img.shields.io/badge/maintainer-Joakim%20Sørensen%20%40ludeeus-blue.svg?style=for-the-badge
 [releases-shield]: https://img.shields.io/github/release/custom-components/blueprint.svg?style=for-the-badge
-[releases]: https://github.com/custom-components/blueprint/releases
+[releases]: https://github.com/custom-components/integration_blueprint/releases
