--- custom_components/blueprint/manifest.json
@@ -1,12 +1,0 @@
-{
-  "domain": "blueprint",
-  "name": "Blueprint",
-  "documentation": "https://github.com/custom-components/blueprint",
-  "issue_tracker": "https://github.com/custom-components/blueprint/issues",
-  "dependencies": [],
-  "config_flow": true,
-  "codeowners": [
-    "@ludeeus"
-  ],
-  "requirements": []
-}
