+++ custom_components/integration_blueprint/entity.py
@@ -4,7 +4,7 @@ from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from .const import DOMAIN, NAME, VERSION, ATTRIBUTION
 
 
-class BlueprintEntity(CoordinatorEntity):
+class IntegrationBlueprintEntity(CoordinatorEntity):
     def __init__(self, coordinator, config_entry):
         super().__init__(coordinator)
         self.config_entry = config_entry
