+++ custom_components/integration_blueprint/manifest.json
@@ -1,0 +1,12 @@
+{
+  "domain": "integration_blueprint",
+  "name": "Integration blueprint",
+  "documentation": "https://github.com/custom-components/integration_blueprint",
+  "issue_tracker": "https://github.com/custom-components/integration_blueprint/issues",
+  "dependencies": [],
+  "config_flow": true,
+  "codeowners": [
+    "@ludeeus"
+  ],
+  "requirements": []
+}
