+++ custom_components/integration_blueprint/translations/en.json
@@ -4,7 +4,7 @@
         "step": {
             "user": {
                 "title": "Blueprint",
-                "description": "If you need help with the configuration have a look here: https://github.com/custom-components/blueprint",
+                "description": "If you need help with the configuration have a look here: https://github.com/custom-components/integration_blueprint",
                 "data": {
                     "username": "Username",
                     "password": "Password"
