+++ custom_components/integration_blueprint/translations/sensor.nb.json
