+++ setup.cfg
@@ -31,5 +31,5 @@ not_skip = __init__.py
 force_sort_within_sections = true
 sections = FUTURE,STDLIB,INBETWEENS,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
 default_section = THIRDPARTY
-known_first_party = custom_components.blueprint 
+known_first_party = custom_components.integration_blueprint
 combine_as_imports = true
